pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sca"
version = "0.1.0"
description = "Semantic Consistency Analyzer — measure LLM output consistency via embedding-space metrics"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "typer>=0.12",
    "textual>=0.60",
    "litellm>=1.40",
    "sentence-transformers>=3.0",
    "hdbscan>=0.8",
    "numpy>=1.26",
    "pydantic-settings>=2.0",
    "scikit-learn>=1.4",
    "rich>=13.0",
]

[project.optional-dependencies]
hf = ["transformers>=4.40", "torch>=2.0"]
gguf = ["llama-cpp-python>=0.2"]
nli = []  # already in core via sentence-transformers cross-encoder
umap = ["umap-learn>=0.5"]
export = ["matplotlib>=3.8"]
all = ["sca[hf,gguf,umap,export]"]

[project.scripts]
sca = "sca.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["sca"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
